requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
apache-beam==2.44.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
azure-core==1.26.3
azure-identity==1.12.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
commonmark==0.9.1
constantly==15.1.0
contourpy==1.0.6
crcmod==1.7
cryptography==39.0.2
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.32
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
Django==4.0.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-oauth-toolkit==2.2.0
django-storages==1.13
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.19
exceptiongroup==1.0.1
fastavro==1.7.1
fasteners==0.18
filelock==3.9.0
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
flatbuffers==22.11.23
fonttools==4.38.0
frozenlist==1.3.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-secret-manager==2.12.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.50.0
grpcio-status==1.48.2
guesser-game==0.1.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.7.0
hdfs==2.7.0
hpack==4.0.0
httpcore==0.16.3
httpie==3.2.1
httplib2==0.20.4
httpx==0.23.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jaraco.classes==3.2.3
jax==0.4.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jwcrypto==1.4.2
keras==2.11.0
keras-nightly==2.13.0.dev2023021208
keyring==23.11.0
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.6.2
microsoft-kiota-abstractions==0.5.0
microsoft-kiota-authentication-azure==0.2.0
microsoft-kiota-http==0.4.0
microsoft-kiota-serialization-json==0.3.0
microsoft-kiota-serialization-text==0.2.0
miniful==0.0.6
more-itertools==9.0.0
msal==1.21.0
msal-extensions==1.0.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0a9
multidict==6.0.2
numpy==1.22.4
oauthlib==3.2.2
objsize==0.6.1
opt-einsum==3.3.0
orjson==3.8.6
outcome==1.2.0
packaging==21.3
pandas==1.5.3
parsel==1.7.0
peupasswd==1.0.1.2
Pillow==9.3.0
pkginfo==1.8.3
portalocker==2.7.0
Protego==0.2.1
proto-plus==1.22.2
protobuf==4.21.12
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pydot==1.4.2
pyFUME==0.2.25
Pygments==2.13.0
PyJWT==2.6.0
pymongo==3.13.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==305
pywin32-ctypes==0.2.0
queuelib==1.6.2
readme-renderer==37.3
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
scapy==2.4.5
scikit-learn==1.2.1
scipy==1.10.0
Scrapy==2.8.0
selenium==4.6.0
service-identity==21.1.0
simpful==2.9.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.2
tb-nightly==2.13.0a20230212
tensorboard==2.11.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tf-estimator-nightly==2.13.0.dev2023021209
tf-nightly==2.13.0.dev20230212
tf_nightly_intel==2.13.0.dev20230212
threadpoolctl==3.1.0
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.9.2
twine==4.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
w3lib==2.1.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.10.0
zope.interface==5.5.2
zstandard==0.19.0
